AQI_PREDICTOR/requirements.txt
streamlit
pandas
joblib
numpy
Pillow
scikit-learn
matplotlib
seaborn
kagglehub
